# Reduce allocations in MediaTokenService and improve performance