--- a/src/OrchardCore.Modules/OrchardCore.Media/Processing/ImageVersionProcessor.cs
+++ b/src/OrchardCore.Modules/OrchardCore.Media/Processing/ImageVersionProcessor.cs
@@ -17,8 +17,10 @@ namespace OrchardCore.Media.Processing
         /// </summary>
         public const string VersionCommand = "v";
 
+        private static readonly IEnumerable<string> VersionCommands = new[] { VersionCommand };
+
         public IEnumerable<string> Commands
-            => new[] { VersionCommand };
+            => VersionCommands;
 
         public FormattedImage Process(FormattedImage image, ILogger logger, IDictionary<string, string> commands, CommandParser parser, CultureInfo culture)
             => image;