--- a/src/OrchardCore.Modules/OrchardCore.Media/Processing/MediaTokenService.cs
+++ b/src/OrchardCore.Modules/OrchardCore.Media/Processing/MediaTokenService.cs
@@ -2,10 +2,12 @@ using System;
 using System.Collections.Generic;
 using System.Security.Cryptography;
 using System.Text;
+using System.Text.Encodings.Web;
 using Cysharp.Text;
 using Microsoft.AspNetCore.WebUtilities;
 using Microsoft.Extensions.Caching.Memory;
 using Microsoft.Extensions.Options;
+using Microsoft.Extensions.Primitives;
 using SixLabors.ImageSharp.Web.Processors;
 
 namespace OrchardCore.Media.Processing
@@ -15,7 +17,7 @@ namespace OrchardCore.Media.Processing
         private const string TokenCacheKeyPrefix = "MediaToken:";
         private readonly IMemoryCache _memoryCache;
 
-        private readonly HashSet<string> _knownCommands = new HashSet<string>();
+        private readonly HashSet<string> _knownCommands = new HashSet<string>(12);
         private readonly byte[] _hashKey;
 
         public MediaTokenService(
@@ -24,6 +26,7 @@ namespace OrchardCore.Media.Processing
             IEnumerable<IImageWebProcessor> processors)
         {
             _memoryCache = memoryCache;
+
             foreach (var processor in processors)
             {
                 foreach (var command in processor.Commands)
@@ -39,34 +42,24 @@ namespace OrchardCore.Media.Processing
         {
             var pathIndex = path.IndexOf('?');
 
-            var parsed = pathIndex != -1
-                ? QueryHelpers.ParseQuery(path.Substring(pathIndex + 1))
-                : null;
+            Dictionary<string, StringValues> processingCommands = null;
+            Dictionary<string, StringValues> otherCommands = null;
+
+            if (pathIndex != -1)
+            {
+                ParseQuery(path.Substring(pathIndex + 1), out processingCommands, out otherCommands);
+            }
 
             // If no commands or only a version command don't bother tokenizing.
-            if (parsed is null || parsed.Count == 0 || parsed.Count == 1 && parsed.ContainsKey(ImageVersionProcessor.VersionCommand))
+            if (processingCommands is null
+                || processingCommands.Count == 0
+                || processingCommands.Count == 1 && processingCommands.ContainsKey(ImageVersionProcessor.VersionCommand))
             {
                 return path;
             }
 
-            var processingCommands = new Dictionary<string, string>(parsed.Count);
-            Dictionary<string, string> otherCommands = null;
-
-            foreach (var command in parsed)
-            {
-                if (_knownCommands.Contains(command.Key))
-                {
-                    processingCommands[command.Key] = command.Value.ToString();
-                }
-                else
-                {
-                    otherCommands ??= new Dictionary<string, string>();
-                    otherCommands[command.Key] = command.Value.ToString();
-                }
-            }
-
             // Using the command values as a key retrieve from cache
-            var queryStringTokenKey = CreateQueryStringTokenKey(processingCommands.Values);
+            var queryStringTokenKey = CreateQueryStringTokenKey(processingCommands);
             var queryStringToken = GetHash(queryStringTokenKey);
 
             processingCommands["token"] = queryStringToken;
@@ -80,12 +73,75 @@ namespace OrchardCore.Media.Processing
                 }
             }
 
-            return QueryHelpers.AddQueryString(path.Substring(0, pathIndex), processingCommands);
+            return AddQueryString(path.AsSpan(0, pathIndex), processingCommands);
         }
+
+#if NET6_0_OR_GREATER // fast version with structs
+        private void ParseQuery(
+            string queryString,
+            out Dictionary<string, StringValues> processingCommands,
+            out Dictionary<string, StringValues> otherCommands)
+        {
+            processingCommands = null;
+            otherCommands = null;
+
+            var accumulator = new KeyValueAccumulator();
+            var otherCommandAccumulator = new KeyValueAccumulator();
+            var enumerable = new QueryStringEnumerable(queryString);
+
+            foreach (var pair in enumerable)
+            {
+                var key = pair.DecodeName().ToString();
+                var value = pair.DecodeValue().ToString();
+
+                if (_knownCommands.Contains(key))
+                {
+                    accumulator.Append(key, value);
+                }
+                else
+                {
+                    otherCommandAccumulator.Append(key, value);
+                }
+            }
+
+            if (accumulator.HasValues)
+            {
+                processingCommands = accumulator.GetResults();
+            }
+
+            if (otherCommandAccumulator.HasValues)
+            {
+                otherCommands = otherCommandAccumulator.GetResults();
+            }
+        }
+#else // slower version until NET 6 is the default
+        private void ParseQuery(
+            string queryString,
+            out Dictionary<string, StringValues> processingCommands,
+            out Dictionary<string, StringValues> otherCommands)
+        {
+            var parsed = QueryHelpers.ParseQuery(queryString);
+            processingCommands = new Dictionary<string, StringValues>();
+            otherCommands = null;
+
+            foreach (var command in parsed)
+            {
+                if (_knownCommands.Contains(command.Key))
+                {
+                    processingCommands[command.Key] = command.Value;
+                }
+                else
+                {
+                    otherCommands ??= new Dictionary<string, StringValues>();
+                    otherCommands[command.Key] = command.Value;
+                }
+            }
+        }
+#endif
 
         public bool TryValidateToken(IDictionary<string, string> commands, string token)
         {
-            var queryStringTokenKey = CreateQueryStringTokenKey(commands.Values);
+            var queryStringTokenKey = CreateQueryStringTokenKey(commands);
 
             // Store a hash of the valid query string commands.
             var queryStringToken = GetHash(queryStringTokenKey);
@@ -98,30 +154,97 @@ namespace OrchardCore.Media.Processing
             return false;
         }
 
-        private static string CreateQueryStringTokenKey(IEnumerable<string> values)
+        /// <summary>
+        /// Specialized version with fast enumeration and no StringValues string allocation.
+        /// </summary>
+        private static string CreateQueryStringTokenKey(Dictionary<string, StringValues> values)
         {
             using var builder = ZString.CreateStringBuilder();
             builder.Append(TokenCacheKeyPrefix);
-            foreach (var item in values)
+            foreach (var pair in values)
             {
-                builder.Append(item);
+                builder.Append(pair.Value.ToString());
+            }
+            return builder.ToString();
+        }
+
+        private static string CreateQueryStringTokenKey(IDictionary<string, string> values)
+        {
+            using var builder = ZString.CreateStringBuilder();
+            builder.Append(TokenCacheKeyPrefix);
+            foreach (var pair in values)
+            {
+                builder.Append(pair.Value);
             }
             return builder.ToString();
         }
 
         private string GetHash(string queryStringTokenKey)
-            => _memoryCache.GetOrCreate(queryStringTokenKey, entry =>
-               {
-                   entry.SlidingExpiration = TimeSpan.FromHours(5);
+        {
+            if (!_memoryCache.TryGetValue(queryStringTokenKey, out var result))
+            {
+                using var entry = _memoryCache.CreateEntry(queryStringTokenKey);
 
-                   using var hmac = new HMACSHA256(_hashKey);
+                entry.SlidingExpiration = TimeSpan.FromHours(5);
 
-                   // queryStringTokenKey also contains prefix
-                   var bytes = Encoding.UTF8.GetBytes(queryStringTokenKey, TokenCacheKeyPrefix.Length, queryStringTokenKey.Length - TokenCacheKeyPrefix.Length);
+                using var hmac = new HMACSHA256(_hashKey);
 
-                   var hash = hmac.ComputeHash(bytes);
+                // queryStringTokenKey also contains prefix
+                var chars = queryStringTokenKey.AsSpan(TokenCacheKeyPrefix.Length);
 
-                   return Convert.ToBase64String(hash);
-               });
+                // only allocate on stack if it's small enough
+                var requiredLength = Encoding.UTF8.GetByteCount(chars);
+                var stringBytes = requiredLength < 1024
+                    ? stackalloc byte[requiredLength]
+                    : new byte[requiredLength];
+
+                // 256 for SHA-256, fits in stack nicely
+                Span<byte> hashBytes = stackalloc byte[hmac.HashSize];
+
+                var stringBytesLength = Encoding.UTF8.GetBytes(chars, stringBytes);
+
+                hmac.TryComputeHash(stringBytes.Slice(0, stringBytesLength), hashBytes, out var hashBytesLength);
+
+                entry.Value = result = Convert.ToBase64String(hashBytes.Slice(0, hashBytesLength));
+            }
+
+            return (string) result;
+        }
+
+        /// <summary>
+        /// Custom version of <see cref="QueryHelpers.AddQueryString(string,string,string)"/> that takes our pre-built
+        /// dictionary, uri as ReadOnlySpan&lt;char&gt; and uses ZString. Otherwise same logic.
+        /// </summary>
+        private static string AddQueryString(
+            ReadOnlySpan<char> uri,
+            Dictionary<string, StringValues> queryString)
+        {
+            var anchorIndex = uri.IndexOf('#');
+            var uriToBeAppended = uri;
+            var anchorText = ReadOnlySpan<char>.Empty;
+            // If there is an anchor, then the query string must be inserted before its first occurrence.
+            if (anchorIndex != -1)
+            {
+                anchorText = uri.Slice(anchorIndex);
+                uriToBeAppended = uri.Slice(0, anchorIndex);
+            }
+
+            var queryIndex = uriToBeAppended.IndexOf('?');
+            var hasQuery = queryIndex != -1;
+
+            using var sb = ZString.CreateStringBuilder();
+            sb.Append(uriToBeAppended);
+            foreach (var parameter in queryString)
+            {
+                sb.Append(hasQuery ? '&' : '?');
+                sb.Append(UrlEncoder.Default.Encode(parameter.Key));
+                sb.Append('=');
+                sb.Append(UrlEncoder.Default.Encode(parameter.Value));
+                hasQuery = true;
+            }
+
+            sb.Append(anchorText);
+            return sb.ToString();
+        }
     }
 }