--- a/src/OrchardCore.Modules/OrchardCore.Media/Processing/TokenCommandProcessor.cs
+++ b/src/OrchardCore.Modules/OrchardCore.Media/Processing/TokenCommandProcessor.cs
@@ -17,8 +17,10 @@ namespace OrchardCore.Media.Processing
         /// </summary>
         public const string TokenCommand = "token";
 
+        private static readonly IEnumerable<string> TokenCommands = new[] { TokenCommand };
+
         public IEnumerable<string> Commands
-            => new[] { TokenCommand };
+            => TokenCommands;
 
         public FormattedImage Process(FormattedImage image, ILogger logger, IDictionary<string, string> commands, CommandParser parser, CultureInfo culture)
             => image;