--- a/test/OrchardCore.Tests/Modules/OrchardCore.Media/MediaTokenTests.cs
+++ b/test/OrchardCore.Tests/Modules/OrchardCore.Media/MediaTokenTests.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Security.Cryptography;
 using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
@@ -16,24 +15,35 @@ namespace OrchardCore.Tests.Modules.OrchardCore.Media
     public class MediaTokenTests
     {
         private readonly MediaTokenSettings _mediaTokenSettings;
+
+        // use static value for repeatable tests
+        private readonly byte[] _hashKey =
+        {
+            88, 204, 124, 2, 54, 72, 44, 8, 207, 85, 88, 9, 245, 129, 200, 41, 13, 169, 54, 213, 202, 118, 197, 9, 26,
+            20, 108, 197, 123, 168, 140, 79, 32, 184, 144, 72, 254, 201, 91, 73, 230, 56, 190, 40, 11, 21, 229, 94, 118,
+            120, 84, 75, 174, 181, 168, 94, 30, 191, 169, 41, 222, 12, 239, 106
+        };
+
         public MediaTokenTests()
         {
-            var rng = RandomNumberGenerator.Create();
             _mediaTokenSettings = new MediaTokenSettings();
-            _mediaTokenSettings.HashKey = new byte[64];
-            rng.GetBytes(_mediaTokenSettings.HashKey);
+            _mediaTokenSettings.HashKey = _hashKey;
         }
 
-        [Fact]
-        public void ShouldAddToken()
+        [Theory]
+        [InlineData("/media/blog.jpg?width=100&height=100", "/media/blog.jpg?width=100&height=100&token=eVmmj09NoysPASEiuhCuUHJR%2BSrUtSafBo738SuL2eU%3D")]
+        [InlineData("/media/blog.jpg?width=100&height=100&foo=bar", "/media/blog.jpg?width=100&height=100&token=eVmmj09NoysPASEiuhCuUHJR%2BSrUtSafBo738SuL2eU%3D&foo=bar")]
+        public void ShouldAddToken(string path, string expected)
         {
             var serviceProvider = CreateServiceProvider();
             var mediaTokenService = serviceProvider.GetRequiredService<IMediaTokenService>();
-            var path = "/media/blog.jpg?width=100&height=100";
 
-            var tokenizedPath = mediaTokenService.AddTokenToPath(path);
-
-            Assert.True(tokenizedPath.Contains("token", StringComparison.OrdinalIgnoreCase));
+            // make sure we also hit cache
+            for (var i = 0; i < 2; ++i)
+            {
+                var tokenizedPath = mediaTokenService.AddTokenToPath(path);
+                Assert.Equal(expected, tokenizedPath);
+            }
         }
 
         [Theory]