One has to have hobbies, right? Felt like a fun exercise as Orchard can use all .NET 6 goodness. 

`MediaTokenService` does some unnecessary allocations and is called on every request which has images served.

* ~~tweak constructor to iterate array, DI gives array and it's fastest to traverse - constructor takes considerable amount of time (see scope note below)~~
  * no need after changing service to singleton, shaved off extra 200ns without the constructor invoke
* don't `ToString` `StringValues`, keep them as-is as in dictionary once built
* parse commands and others commands with one pass using constructs that `QueryHelpers` uses
* allocate static command arrays to return from processors, some of them were always returning new ones
* change `GetHash` not to allocation anonymous closure each time by using `Try`/`Set `construct, uses same logic that extension method `GetOrCreate` uses
* change hash calculation to use stackalloc and spans for workloads < 1024 bytes (common case)
* use custom `AddQueryString `method that understands spans and concrete dictionary type

`MediaTokenService` seems to be scoped service and thus known commands need to be initialized for every request, is this a tenancy things and cannot be made singleton?

## Benchmarks

``` ini

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.22000
AMD Ryzen 9 5950X, 1 CPU, 32 logical and 16 physical cores
.NET SDK=6.0.101
  [Host]     : .NET 6.0.1 (6.0.121.56705), X64 RyuJIT
  DefaultJob : .NET 6.0.1 (6.0.121.56705), X64 RyuJIT


```

### Before

|                          Method |     Mean |     Error |    StdDev |  Gen 0 |  Gen 1 | Allocated |
|-------------------------------- |---------:|----------:|----------:|-------:|-------:|----------:|
|                  AddTokenToPath | 1.335 us | 0.0097 us | 0.0091 us | 0.1984 |      - |      3 KB |
|          AddTokenToPath_NoCache | 1.993 us | 0.0089 us | 0.0083 us | 0.2441 |      - |      4 KB |
|         AddTokenToPath_LongPath | 2.474 us | 0.0157 us | 0.0139 us | 0.4005 | 0.0038 |      7 KB |
| AddTokenToPath_LongPath_NoCache | 3.135 us | 0.0098 us | 0.0091 us | 0.4578 | 0.0038 |      7 KB |

### After

Now testing with singleton - unlike main branch!

|                          Method |       Mean |   Error |  StdDev |  Gen 0 | Allocated |
|-------------------------------- |-----------:|--------:|--------:|-------:|----------:|
|                  AddTokenToPath |   696.1 ns | 1.18 ns | 1.05 ns | 0.0715 |      1 KB |
|          AddTokenToPath_NoCache | 1,309.6 ns | 2.28 ns | 1.91 ns | 0.1087 |      2 KB |
|         AddTokenToPath_LongPath | 1,556.8 ns | 4.93 ns | 4.37 ns | 0.2060 |      3 KB |
| AddTokenToPath_LongPath_NoCache | 2,248.9 ns | 6.70 ns | 6.27 ns | 0.2441 |      4 KB |