@deanmarcussen 2022-01-02
> One has to have hobbies, right? Felt like a fun exercise as Orchard can use all .NET 6 goodness.

Indeed :)

We aren't quite there with .NET 6 only yet, still multi targetting. Due to drop 3.1 and 5 after the next release (due shortly)

> `MediaTokenService` seems to be scoped service and thus known commands need to be initialized for every request, is this a tenancy things and cannot be made singleton?

No, it's just Scoped until you find value in making it a Singleton.