@lahma 2022-01-02
I tweaked existing test case to test both with known and unknown commands and it now checks for complete generated string, results didn't change between main and this branch.